requirements.txt
google-cloud-documentai
google-cloud-storage
google-cloud-documentai-toolbox
grpcio
dotenv
langchain~=0.0.171
openai~=0.27.6
tiktoken
gradio
python-dotenv~=1.0.0
fastapi~=0.95.1
pydantic~=1.10.7
numpy~=1.24.3